fastapi-mongodb/pyproject.toml
[project]
name = "fastapi-mongodb"
dynamic = ["version"]
readme = "readme.md"
license = {text = "MIT NON-AI License"}
authors = [
    {name = "Rajan Khullar", email = "rkhullar03@gmail.com"}
]
requires-python = ">= 3.8"
dependencies = [
    "httpx",
    "fastapi",
    "pydantic-settings"
]

[project.optional-dependencies]
server = ["uvicorn"]
lambda = ["mangum"]
mongodb = ["pymongo[aws,srv]"]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"
